List, set, dict and tuple all allow aligned_indent
List, set, dict and tuple all use aligned_indent which permits both hanging and aligned styles.

The final closing delim can be optionally specified to dedent (python says both are fine, allow queries to specify which to use).

Multiple opening delims on one line and multiple closing on a line are collapsed so as not to over indent,

The final line of aligned_indent blocks which must in some cases be treated specially to avoid clashing with the next line is treated the same regardless of whether the aligned_indent node actually uses aligned indentation or just normal indentation.

Finally, try: on it’s own will indent when typing live but make no guaranteeds about whole-file formatting.